+++ src/main/java/io/anymobi/repositories/jpa/security/GroupsRepository.java
@@ -3,7 +3,7 @@ package io.anymobi.repositories.jpa.security;
 import io.anymobi.domain.entity.sec.Groups;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupsRepository extends JpaRepository<Groups, Integer> {
+public interface GroupsRepository extends JpaRepository<Groups, Long> {
 
     Groups findByGroupName(String name);
