+++ src/main/java/io/anymobi/repositories/jpa/security/ResourcesRepository.java
@@ -3,7 +3,7 @@ package io.anymobi.repositories.jpa.security;
 import io.anymobi.domain.entity.sec.Resources;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResourcesRepository extends JpaRepository<Resources, Integer>{
+public interface ResourcesRepository extends JpaRepository<Resources, Long>{
 
     Resources findByResourceName(String name);
 }
